requirements.txt
streamlit
langchain==0.3.*
langchain-community
langchain-chroma
langchain-huggingface
langchain-groq
pymupdf
sentence-transformers
flashrank
fpdf2
llama-parse